Code Review Completed
Good job! The code is well-organized in appropriate folders and the README is very clear and descriptive!

Here are a few suggestions:
It would be more user-friendly to add a pagination button to navigate through all the accounts more efficiently.
It would be more user-friendly to update the form input fields as mentioned in the previous commit.